requirements.txt
Django==3.2.20
coverage
pyaml
django-recurrence
python-decouple==3.3
pandas
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow>=11.0.0
